% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rug_utils.R
\name{read_data}
\alias{read_data}
\title{Reads columns from a file in table format}
\usage{
read_data(filename, select_columns)
}
\arguments{
\item{filename}{a string data file name including the relative path}

\item{select_columns}{a vector including the column names to be read from the data file}
}
\value{
a data.table object
}
\description{
Additional info
}
