% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rug_utils.R
\name{validate_parameters}
\alias{validate_parameters}
\title{Validate that a JSON file meets a required schema}
\usage{
validate_parameters(jsonparams, jsonschema, cds_package = "rugplot")
}
\arguments{
\item{jsonparams}{A JSON file to be validated against a schema}

\item{jsonschema}{A predefined JSON schema}

\item{cds_package}{The package that contains the JSON schema}
}
\description{
Validate that a JSON file meets a required schema
}
